Splitting the controller into two sub controllers
