+++ runlab/src/main/java/RunLab/EndPoints/Strava/AthleteController.java
@@ -1,4 +1,4 @@
-package RunLab.EndPoints;
+package RunLab.EndPoints.Strava;
 
 import java.io.IOException;
 
@@ -7,44 +7,33 @@ import okhttp3.Response;
 import com.google.gson.Gson;
 
 import RunLab.Exceptions.InvalidRequest;
-import RunLab.Models.*;
+
 import RunLab.Objects.Strava.*;
 import RunLab.Responces.*;
 import RunLab.Wrappers.*;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/Strava/v1")
+@RequestMapping("/strava-api/v1/athlete")
 @CrossOrigin(origins = "http://localhost:3000", maxAge = 3600)
-public class StravaController {
+public class AthleteController {
 
     private Gson gson = new Gson();
-    private Strava stravaWrapper = new Strava();
+    @Autowired
+    private Strava stravaWrapper;
     private Logger logger = LoggerFactory.getLogger(getClass());
 
-    StravaController() {}
+    AthleteController() {}
 
-    @GetMapping("/refresh")
-    public CustomResponse<SummaryActivity[]> refresh() throws InvalidRequest, IOException {
-        logger.info("Refresh Called");
-        Response response = this.stravaWrapper.pull();
-        Success<SummaryActivity[]> r = new Success<SummaryActivity[]>();
-        SummaryActivity[] activity = gson.fromJson(response.body().string().replace("\"", "'"), SummaryActivity[].class);
-        
-        r.setBody(activity);
-        return r;
-    }
-
-    @GetMapping("/getAthleteProfile")
+    @GetMapping("/profile")
     public CustomResponse<AthleteProfile> getAthleteProfile()  throws InvalidRequest, IOException {
         logger.info("Athlete profile Called");
         Response response = this.stravaWrapper.getAthleteProfile();
@@ -55,7 +44,7 @@ public class StravaController {
         return r;
     }
 
-    @GetMapping(value = { "/getAthleteStatistics" })
+    @GetMapping(value = { "/statistics" })
     public CustomResponse<AthleteStatistics> getAthleteStatistics(@RequestParam(value = "ID", defaultValue="16443776") String ID) throws InvalidRequest, IOException {
         logger.info("Athlete Statistics Called");
         Response response = this.stravaWrapper.getAthleteStats(ID);
@@ -66,28 +55,15 @@ public class StravaController {
         return r;
     }
 
-    @GetMapping("/oauth")
-    public CustomResponse<String> oauthGET() {
-        logger.info("Check if Authenticated");
-        // pull in and update data
-        boolean isAuthenticated = this.stravaWrapper.refreshAuthTokens();
-
-        if (isAuthenticated) {
-            return new Success<String>();
-        } else {
-            return new Failure<String>();
-        }
+    @GetMapping("/activies")
+    public CustomResponse<SummaryActivity[]> refresh() throws InvalidRequest, IOException {
+        logger.info("Refresh Called");
+        Response response = this.stravaWrapper.pull();
+        Success<SummaryActivity[]> r = new Success<SummaryActivity[]>();
+        SummaryActivity[] activity = gson.fromJson(response.body().string().replace("\"", "'"), SummaryActivity[].class);
+        
+        r.setBody(activity);
+        return r;
     }
 
-    @PostMapping("/oauth")
-    public CustomResponse<String> oauthPOST(@RequestBody codeModel requestBody) {
-        logger.info("Authentication Called");
-        Boolean dataUpdated = this.stravaWrapper.setAuthTokens(requestBody);
-
-        if (dataUpdated) {
-            return new Success<String>();
-        } else {
-            return new Failure<String>();
-        }
-    }
 }
